(PC-20410) refacto(datetime): extract datetime helpers
